Final Year/requirements.txt
streamlit
mysql.connector
matplotlib.pyplot
hashlib
webbrowser
secrets